+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Revision history for lambdasound
 
-## Unreleased
+## 1.0.1 -- 2023-10-13
 
 * Add `withSampledSound` and `withSampledSoundPulse`
+* Fix incorrect memoization for sequential and parallel sounds
 
 ## 1.0.0 -- 2023-10-12